requirements.txt
Django==2.2.10
django-environ==0.4.5
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
psycopg2==2.8.6
# django-rest-swagger
